+++ src/kataloger/__init__.py
@@ -1,2 +1,2 @@
-__version__: str = "0.3.3"
+__version__: str = "0.3.4"
 package_name: str = "kataloger"